# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_document_documents_upload_post": {
        "properties": {
          "document_type": {
            "$ref": "#/components/schemas/DocumentType",
            "default": "unknown"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "PDF to ingest",
            "title": "File",
            "type": "string"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType",
            "default": "upload"
          },
          "uploaded_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded By"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_document_documents_upload_post",
        "type": "object"
      },
...
      "ChunkType": {
        "enum": [
          "paragraph",
          "table",
          "heading",
          "list",
          "footnote",
          "unknown"
        ],
        "title": "ChunkType",
        "type": "string"
      },
...
      "DocumentChunkRead": {
        "properties": {
          "char_end": {
            "title": "Char End",
            "type": "integer"
          },
          "char_start": {
            "title": "Char Start",
            "type": "integer"
          },
          "chunk_text": {
            "title": "Chunk Text",
            "type": "string"
          },
          "chunk_type": {
            "$ref": "#/components/schemas/ChunkType"
          },
          "fts_config": {
            "title": "Fts Config",
            "type": "string"
          },
          "hash": {
            "title": "Hash",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "ordinal": {
            "title": "Ordinal",
            "type": "integer"
          },
          "page_number": {
            "title": "Page Number",
            "type": "integer"
          },
          "section_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Title"
          }
        },
        "required": [
          "id",
          "page_number",
          "ordinal",
          "section_title",
          "chunk_type",
          "language",
          "fts_config",
          "char_start",
          "char_end",
          "hash",
          "chunk_text"
        ],
        "title": "DocumentChunkRead",
        "type": "object"
      },
      "DocumentPageRead": {
        "properties": {
          "char_count": {
            "title": "Char Count",
            "type": "integer"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "garbled_unicode_ratio": {
            "title": "Garbled Unicode Ratio",
            "type": "number"
          },
          "has_text_layer": {
            "title": "Has Text Layer",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image_area_ratio": {
            "title": "Image Area Ratio",
            "type": "number"
          },
          "page_number": {
            "title": "Page Number",
            "type": "integer"
          },
          "parse_method": {
            "$ref": "#/components/schemas/ParseMethod"
          },
          "vlm_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vlm Reason"
          },
          "vlm_used": {
            "title": "Vlm Used",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "page_number",
          "char_count",
          "image_area_ratio",
          "has_text_layer",
          "garbled_unicode_ratio",
          "parse_method",
          "vlm_used",
          "vlm_reason",
          "confidence"
        ],
        "title": "DocumentPageRead",
        "type": "object"
      },
      "DocumentRead": {
        "properties": {
          "byte_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Byte Size"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "issuer_name_guess": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuer Name Guess"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "parse_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parse Confidence"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "storage_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Uri"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "uploaded_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded By"
          }
        },
        "required": [
          "id",
          "sha256",
          "filename",
          "storage_uri",
          "byte_size",
          "source_type",
          "document_type",
          "status",
          "language",
          "issuer_name_guess",
          "page_count",
          "parse_confidence",
          "uploaded_by",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentRead",
        "type": "object"
      },
      "DocumentStatus": {
        "enum": [
          "uploaded",
          "parsing",
          "parsed",
          "chunked",
          "embedded",
          "extracting",
          "extracted",
          "failed",
          "budget_exceeded"
        ],
        "title": "DocumentStatus",
        "type": "string"
      },
      "DocumentStatusRead": {
        "description": "Where a document has got to, for a client that is watching it.\n\nAssembled rather than reflected: `documents.status` alone cannot say\nwhether the worker has picked the document up, and `extraction_jobs` alone\ncannot say how many pages came out. Both are needed to answer \"is it done,\nand did it work?\", which is the only question this endpoint exists for.",
        "properties": {
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/IngestionJobRead"
            },
            "title": "Jobs",
            "type": "array"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "pages_flagged_for_vlm": {
            "title": "Pages Flagged For Vlm",
            "type": "integer"
          },
          "parse_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parse Confidence"
          },
          "searchable": {
            "title": "Searchable",
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "terminal": {
            "title": "Terminal",
            "type": "boolean"
          }
        },
        "required": [
          "document_id",
          "filename",
          "status",
          "searchable",
          "page_count",
          "chunk_count",
          "pages_flagged_for_vlm",
          "parse_confidence",
          "jobs",
          "terminal"
        ],
        "title": "DocumentStatusRead",
        "type": "object"
      },
      "DocumentType": {
        "enum": [
          "prospectus",
          "information_memorandum",
          "trust_deed",
          "rating_report",
          "announcement",
          "supplemental",
          "unknown"
        ],
        "title": "DocumentType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestionResponse": {
        "properties": {
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "page_count": {
            "title": "Page Count",
            "type": "integer"
          },
          "pages_flagged_for_vlm": {
            "title": "Pages Flagged For Vlm",
            "type": "integer"
          },
          "skipped": {
            "title": "Skipped",
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          }
        },
        "required": [
          "document_id",
          "status",
          "page_count",
          "chunk_count",
          "pages_flagged_for_vlm",
          "skipped"
        ],
        "title": "IngestionResponse",
        "type": "object"
      },
...
      "Language": {
        "enum": [
          "en",
          "ms",
          "mixed",
          "unknown"
        ],
        "title": "Language",
        "type": "string"
      },
...
      "ParseMethod": {
        "enum": [
          "pymupdf",
          "pdfplumber",
          "vlm",
          "none"
        ],
        "title": "ParseMethod",
        "type": "string"
      },
...
      "SourceType": {
        "enum": [
          "upload",
          "email",
          "bursa",
          "rating_agency",
          "internal"
        ],
        "title": "SourceType",
        "type": "string"
      },
...
      "UploadResponse": {
        "description": "Upload outcome.\n\n`duplicate` is a normal result, not an error: the same bytes uploaded twice\nare one document (CLAUDE.md 1.7), and the caller gets the existing row back.",
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentRead"
          },
          "duplicate": {
            "title": "Duplicate",
            "type": "boolean"
          }
        },
        "required": [
          "duplicate",
          "document"
        ],
        "title": "UploadResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Sukuk & bond covenant intelligence platform",
    "title": "OpusCovIntel",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/documents": {
      "get": {
        "operationId": "list_documents_documents_get",
        "parameters": [
          {
            "in": "query",
            "name": "status_filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DocumentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentRead"
                  },
                  "title": "Response List Documents Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List documents",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/upload": {
      "post": {
        "description": "Store the PDF and queue it for parsing.\n\nReturns 201 for a new document and 200 when the bytes were already known --\na duplicate is a correct outcome, not an error (CLAUDE.md 1.7).",
        "operationId": "upload_document_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload a PDF for ingestion",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{document_id}": {
      "get": {
        "operationId": "get_document_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get one document",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{document_id}/chunks": {
      "get": {
        "operationId": "list_chunks_documents__document_id__chunks_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentChunkRead"
                  },
                  "title": "Response List Chunks Documents  Document Id  Chunks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chunks with their character spans",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/documents/{document_id}/pages": {
      "get": {
        "operationId": "list_pages_documents__document_id__pages_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentPageRead"
                  },
                  "title": "Response List Pages Documents  Document Id  Pages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Per-page parse telemetry",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{document_id}/process": {
      "post": {
        "description": "Run ingestion inline, including indexing.\n\nThe worker picks queued documents up on its own; this exists for operators\nand demos that do not want to wait for a poll interval.\n\nIndexing is part of it. A document ingested through this endpoint has no\nqueued parse job left for the worker to claim, so if this path stopped at\nchunked the document would never become searchable at all.",
        "operationId": "process_document_documents__document_id__process_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Parse and chunk now, without waiting for the worker",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{document_id}/status": {
      "get": {
        "description": "Progress for a client that is watching an upload.\n\nThe worker has always recorded this -- `extraction_jobs` is the queue, and\nit carries the status, the timings and the failure message -- but nothing\nexposed it over HTTP, so the only way to know whether an upload had been\ningested was to look in the database (docs/review.md, finding 7).",
        "operationId": "document_status_documents__document_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentStatusRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Where a document has got to in the pipeline",
        "tags": [
          "documents"
        ]
      }
    },
...
  }
}